api/requirements.txt
Django==5.0
djangorestframework==3.14.0
pymongo==4.6.1
django-cors-headers==4.3.1
